% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bas_plot}
\alias{bas_plot}
\title{Beautified version of summary plot for BAS}
\usage{
bas_plot(x, top.models = 20)
}
\arguments{
\item{x}{A BAS model object.}

\item{top.models}{The number of top models to include in the plot.}
}
\value{
A gtable.
}
\description{
Beautified version of summary plot for BAS
}
