% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bp_distance}
\alias{bp_distance}
\title{Calculate the distance between to break points}
\usage{
bp_distance(s1, e1, s2, e2)
}
\arguments{
\item{s1}{Start point of break point 1}

\item{e1}{End point of break point 1}

\item{s2}{Start point of break point 2}

\item{e2}{End point of break point 2}
}
\value{
A numeric value for the distance between the two break points.
}
\description{
Calculate the distance between to break points
}
