% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bp_distance_check}
\alias{bp_distance_check}
\title{Check that the two break points are sane, in that one does not fall exclusively within the other}
\usage{
bp_distance_check(s1, e1, s2, e2)
}
\arguments{
\item{s1}{Start point of break point 1}

\item{e1}{End point of break point 1}

\item{s2}{Start point of break point 2}

\item{e2}{End point of break point 2}
}
\value{
Logical. TRUE if the the two break points are sane, FALSE if not.
}
\description{
Check that the two break points are sane, in that one does not fall exclusively within the other
}
