% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci-lianti.R
\name{Cell}
\alias{Cell}
\title{The Cell object}
\usage{
Cell(vcf_file)
}
\arguments{
\item{vcf_file}{The path to the input vcf file.}
}
\value{
A Cell object that is a list.
}
\description{
The Cell object that stores information about one cell.
}
