% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_cell_pca}
\alias{plot_cell_pca}
\title{Color cell attributes on scatter plot of the first two PCs}
\usage{
plot_cell_pca(cell_pca, cell_df)
}
\arguments{
\item{cell_pca}{The PCA object.}

\item{cell_df}{The data frame of cell attributes.}
}
\value{
a list of ggplots.
}
\description{
Color cell attributes on scatter plot of the first two PCs
}
