% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci-lianti.R
\name{restoreSciLianti}
\alias{restoreSciLianti}
\title{Restore a trimmed Cell object to full size}
\usage{
restoreSciLianti(cell, bin_window_size = NA)
}
\arguments{
\item{cell}{A trimmed Cell object.}

\item{bin_window_size}{The window size to calculate binned allele frequencies. Will use stored
old value if left as NA.}
}
\value{
A full-size Cell object.
}
\description{
Restore a trimmed Cell object to full size
}
\examples{
\dontrun{
cell <- restoreSciLianti(cell, bin_window_size=NA)
}

}
