% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci-lianti.R
\name{sciLianti}
\alias{sciLianti}
\title{Analyze break points for one cell}
\usage{
sciLianti(single_cell_file, bin_window_size, filter = "none",
  organism = "mouse", known_vcf = NA)
}
\arguments{
\item{single_cell_file}{Path to the input vcf file.}

\item{bin_window_size}{Window size for calculating binned allele frequencies.}

\item{filter}{Filter SNPs based on annotation. One of c('none', 'rs', 'rslow').}

\item{organism}{One of c('mouse', 'fly').}

\item{known_vcf}{Path to a vcf file containing known SNPs. Only used when organism is 'fly' and when we
want to filter by known SNPs.}
}
\value{
A Cell object.
}
\description{
Analyze break points for one cell.
}
\examples{
\dontrun{
cell <- sciLianti(single_cell_file="/path/to/input/vcf", bin_window_size=40, filter='rslow',
 organism='mouse', known_vcf=NA)
}

}
