% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slfile}
\alias{slfile}
\title{Locate a file installed with this library}
\usage{
slfile(f)
}
\arguments{
\item{f}{The name of a file installed with this library.}
}
\value{
The reconstructed full file path of said file.
}
\description{
Locate a file installed with this library
}
