% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci-lianti.R
\name{trimSciLianti}
\alias{trimSciLianti}
\title{Make a Cell object smaller in size by removing data that can be relatively easily regenerated,
for storage purpose}
\usage{
trimSciLianti(cell)
}
\arguments{
\item{cell}{A processed full-size Cell object.}
}
\value{
A trimmed Cell object.
}
\description{
Make a Cell object smaller in size by removing data that can be relatively easily regenerated,
for storage purpose
}
\examples{
\dontrun{
cell <- trimSciLianti(cell)
}

}
